Bug fix: duplicate key error on adding new row.
For tables imported through ogr2ogr, whenever a new row was added,
"duplicate key" error was being raised for the primary key of the table.
Current value of sequence was not being updated to take into account
the maximum value of the primary key in the table.
This did not happen earlier. May be something in ogr2ogr or related
dependency changed.